Simplify shutting down workflow
We renamed state.draining? to state.shutting_down? to
avoid confusion with draining while revoking.

Also because the revoke draining would continue to
schedule poll messages, but not poll them, we changed
the shut down draining to do the same. Now scheduling
always continues, but there is a single check, guarding
the poll function, that knows if the poll should
effectively happen.

Unifying the flows help us avoid complexity when it
comes to understanding what happens when both revoke
and draining are happening at the same time.