push read item to the end of queue (#110)
This way we are basically implementing a ring buffer. After the data
for an specific partition is read, we put it back to end of queue. This
allows fairer consumption of cross partitions.

Before this patch, we always read from a single partitions and then
move on the next partitions. However since we are using
GenStage.PartitionDispatcher (using partition as partition key),
the messages get stuck into the dispatcher pending for a single
consumer to consume the data.